fixxing

+++ src/b_plus_tree.cpp
@@ -4,9 +4,9 @@ BPlusTree::BPlusTree(Storage *indexes, Storage *records, size_t nodeSize) {
     // Checking which is the number of keys that a node can contain
     // In the start, the node has a bool and int attributes, so we subtract them from the nodeSize
     // For n keys, we have n+1 pointers, so we subtract one Address too, then we try increasing the key number
-    // with having that for every key there is a float (the actual key) and the address of the record
+    // with having that for every key there is an int (the actual key) and the address of the record
     maxNumKeys = 0;
-    while (((maxNumKeys + 1) * (sizeof(Address) + sizeof(float)) + sizeof(int) + sizeof(bool) + sizeof(Address)) <
+    while (((maxNumKeys + 1) * (sizeof(Address) + sizeof(int)) + sizeof(int) + sizeof(bool) + sizeof(Address)) <
            nodeSize) {
         maxNumKeys += 1;
     }
